store comments in Datastore
