+++ portfolio/pom.xml
@@ -28,8 +28,12 @@
       <artifactId>gson</artifactId>
       <version>2.8.6</version>
     </dependency>
+    <dependency>
+      <groupId>com.google.appengine</groupId>
+      <artifactId>appengine-api-1.0-sdk</artifactId>
+      <version>1.9.59</version>
+      </dependency>
   </dependencies>
- 
   <build>
     <plugins>
       <!-- Provides `mvn package appengine:run` for local testing
