+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -12,34 +12,10 @@ public class Comment {
     private String comment;
     private String date;
 
-    public Comment(){
-        name="";
-        comment="";
-        date="";
-    }
     public Comment(String n, String c){
         name = n;
         comment = c;
         date = getCurrentDate();
-    }
-
-    public void setName(String n){
-        name = n;
-    }
-    public void setComment(String c){
-        comment = c;
-    }
-    public void setDate(){
-      date = getCurrentDate();
-    }
-    public String getName(){
-        return name;
-    }
-    public String getComment(){
-        return comment;
-    }
-    public String getDate() {
-        return date;
     }
 
     private String getCurrentDate() {
@@ -47,5 +23,9 @@ public class Comment {
       LocalDateTime now = LocalDateTime.now();  
       String output = dtf.format(now);
       return output;
+    }
+
+    public String getDate(){
+        return date;
     }
 }
