+++ portfolio/src/main/java/com/google/sps/servlets/NewCommentServlet.java
@@ -1,0 +1,67 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.sps.servlets;
+
+
+import com.google.appengine.api.datastore.DatastoreService;
+import com.google.appengine.api.datastore.DatastoreServiceFactory;
+import com.google.appengine.api.datastore.Entity;
+import com.google.sps.data.Comment;
+import com.google.gson.Gson;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/** Servlet that returns some example content. TODO: modify this file to handle comments data */
+@WebServlet("/new-comment")
+public class NewCommentServlet extends HttpServlet {
+
+  @Override
+  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    
+    String name = request.getParameter("name");
+    //TODO: make a min character requirement for name
+    if(name.length() < 4) {
+      System.out.println("Name too short");
+      response.setContentType("text/html");
+      response.getWriter().println("Please enter a name that has at least 3 characters");
+    }
+
+    String text = request.getParameter("comment-box");
+    //TODO: make a min character requirement for text
+    if(text.length() < 4) {
+      System.out.println("Comment too short");
+      response.setContentType("text/html");
+      response.getWriter().println("Please enter a comment that has at least 3 characters");
+    }
+
+    Comment comment = new Comment(name,text);
+    Entity com = new Entity("Comment");
+    com.setProperty("name", name);
+    com.setProperty("comment", text);
+    com.setProperty("date", comment.getDate());
+
+    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+    datastore.put(com);
+    
+    response.sendRedirect("pages/secret-talents.html");
+  }
+
+
+}
