+++ portfolio/src/main/webapp/pages/secret-talents.html
@@ -14,7 +14,7 @@
       <li class="right"><a href="gallery.html">GALLERY</a></li>
       <li class="right"><a href="../index.html">HOME</a></li>
     </ul>
-    <div class="talents-wrap bigger">
+    <div class="talents-wrap smaller">
       <div class="talents-nowrap">
         <div class="column50">
           <img class="talent" src="../images/talents/beekeeper.jpg"/ alt="Isabel in a beekeeper's hat and veil.">
@@ -38,13 +38,13 @@
         </ul>
       </div>
       <div class="talent-text">
-        <form action="/comment" method="POST">      
+        <form action="/new-comment" method="POST">      
             <h2>Comment Your Secret Talents Below!</h2><br>
             <label for="name">Name:</label><br>
             <input type="text" id="name" name="name"><br>
             <label for="comment-box"></label><br>
             <textarea id="comment-box" name="comment-box"></textarea>
-            <br>
+            <br><br>
             <input type="submit" onclick="getComments()" value="Submit"><br><br>
         </form> 
       </div>
